# Adding a perf test for Double.ToString("R") and Single.ToString("R")