--- a/tests/src/JIT/Performance/CodeQuality/Numerics/SingleTests.cs
+++ b/tests/src/JIT/Performance/CodeQuality/Numerics/SingleTests.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Numerics
+{
+    public static partial class SingleTests
+    {
+        private const int ToString_RIterations = 5000;
+    }
+}