This adds a basic perf test for `Double.ToString("R")` and `Single.ToString("R")`.

Testing locally, for x64 Windows I see:
```
Running double.tostring_r test...
  Warming up...
    Total Time: 0.0016542
  Executing 1000 iterations...
    Total Time: 1.5293879 seconds
    Average Time: 0.0015293879 seconds
Running single.tostring_r test...
  Warming up...
    Total Time: 0.0011746
  Executing 1000 iterations...
    Total Time: 1.0363284 seconds
    Average Time: 0.0010363284 seconds
```

For x86 Windows I see:
```
Running double.tostring_r test...
  Warming up...
    Total Time: 0.0023679
  Executing 1000 iterations...
    Total Time: 2.2525623 seconds
    Average Time: 0.0022525623 seconds
Running single.tostring_r test...
  Warming up...
    Total Time: 0.001681
  Executing 1000 iterations...
    Total Time: 1.5032813 seconds
    Average Time: 0.0015032813 seconds
```