@tannergooding 2018-09-11
CC. @jkotas, @danmosemsft, @eerhardt