@tannergooding 2018-09-11
@jkotas, I may have just missed it, but I didn't think we had any mechanism which allowed us to run the CoreFX benchmarks against CoreCLR as part of a PR.