@tannergooding 2018-09-11
Nor does it look like we currently support running the CoreFX benchmarks as x86 vs x64 (looking at the metrics on benchview)