@jkotas 2018-09-11
> I didn't think we had any mechanism which allowed us to run the CoreFX benchmarks against CoreCLR as part of a PR.

Right, I wish there was a way to trigger stack correctness or functionality tests in the CI (not just CoreFX - but also ASP.NET or ML.NET), but we are not there yet.